Move persisted form and fix react key error
